requirements.txt
apiclient==1.0.4
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Faker==18.7.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pygraphviz==1.10
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
